Add spread multiplier mod

+++ Modular Weapon System/Assets/Scripts/Entities/Guns/Gun Modifiers/Generic/ProjectileSpreadMod.cs
@@ -4,7 +4,7 @@
 public class ProjectileSpreadMod : GunModifier
 {
     [Tooltip("Bullet spread adjustment as a value before additional modifiers.")]
-    [SerializeField] int projectileSpreadValue;
+    [SerializeField] float projectileSpreadValue;
 
     float ProjectileSpreadRadius(float currentValue){
         return currentValue + projectileSpreadValue;
